+++ CHANGELOG.md
@@ -108,6 +108,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Fixed long labels in IT Hygiene horizontal bar visualizations causing display issues [#8330](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8330)
 - Fixed rendering of the Tactics and Techniques cells in the MITRE ATT&CK flyout [#8516](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8516)
 - Fixed custom filter buttons not being rendered in pdf reports [#8525](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8525)
+- Fixed home severity visualization by removing pinned agent filter [#8551](https://github.com/wazuh/wazuh-dashboard-plugins/pull/8551)
 
 ### Removed